requirements.txt
Flask==3.0.0
anthropic==0.25.0
PyMuPDF==1.23.0
python-pptx==0.6.23
Pillow==10.0.0
markdown2==2.4.10
weasyprint==60.0
python-dotenv==1.0.0
Werkzeug==3.0.0
